feat: file tree organised/structure
