+++ package.json
@@ -11,9 +11,12 @@
     "@types/react": "^18.0.9",
     "@types/react-dom": "^18.0.4",
     "again": "^0.0.1",
+    "cors": "^2.8.5",
     "include-media": "^1.4.10",
+    "node-sass": "^7.0.1",
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
+    "react-router-dom": "^6.3.0",
     "react-scripts": "5.0.1",
     "typescript": "^4.6.4",
     "web-vitals": "^2.1.4"
