+++ src/Nav/Pages/About/About.tsx
@@ -1,0 +1,1 @@
+import React, { useState, useEffect } from "react";
