+++ src/default.scss
@@ -1,4 +1,4 @@
-$bgColor: rgba(66, 68, 90, 1);
+$bgColor: #282c34;
 $borderRad: 15px;
 $accentColor: #29a165;
 $focusColor: #23a5ad;
