+++ src/index.scss
@@ -8,13 +8,20 @@ body {
   -moz-osx-font-smoothing: grayscale;
   background-color: #282c34;
 }
+
+* {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+}
+
 #root {
   display: flex;
   justify-content: center;
   align-items: center;
   .wrapper {
     display: flex;
-    justify-content: center;
+    flex-direction: column;
     align-items: center;
     width: 100vw;
     height: 100vh;
